+++ nufhe/lwe.py
@@ -40,6 +40,7 @@ from .lwe_gpu import (
     LweLinear,
     LweNoiselessTrivial,
     LweNoiselessTrivialConstant,
+    LweRoll,
     )
 from .random_numbers import (
     rand_uniform_bool,
@@ -176,6 +177,28 @@ class LweSampleArray:
         """
         return LweSampleArray(
             self.params, self.a.copy(), self.b.copy(), self.current_variances.copy())
+
+    def roll(self, shift, axis=None):
+        """
+        Cyclically shifts encrypted bits of the cyphertext **inplace**
+        by ``shift`` positions to the right along ``axis``.
+        ``shift`` can be negative (in which case the elements are shifted to the left).
+        Elements that are shifted beyond the last position are re-introduced at the first
+        (and vice versa).
+        If ``axis`` is ``None``, the last axis is used.
+
+        Works equivalently to ``numpy.roll``.
+        """
+        if shift == 0:
+            return
+
+        if axis is None:
+            axis = len(self.shape) - 1
+        else:
+            axis = axis % len(self.shape)
+
+        comp = get_computation(self.a.thread, LweRoll, self.shape_info, axis)
+        comp(self.a, self.b, self.current_variances, shift)
 
     def dump(self, file_obj):
         """